Init , the first version
allow select between two coordinates and know what is the width and height of selected area. allow width control + c copy the information. work fine in cinammon you can pin the icon to the bottom bar.  a bug is you need maximize the icon app for works. the app was was written in vala.  Depends: libgtk-3-0, desktop-file-utils
 . Enjoy